devops/scripts/requirements.txt
# DevOps Learning Hub - Content Generator Dependencies
google-generativeai>=0.4.0
python-dotenv>=1.0.0
